+++ packages/cli/src/index.test.ts
