+++ packages/schema-to-library/src/helper/dep/resolve-schema-dependencies-from-shema.ts
@@ -1,4 +1,4 @@
-import type { Schema } from '../../cli/index.js'
+import type { Schema } from '../../../../cli/src/index.js'
 
 export function resolveSchemaDependenciesFromSchema(schema: Schema): string[] {
   // Merge both definitions and $defs
